Update main.css
added box models for each section and each image. The padding, border, and margin were fixed for each image and section to match. I also made each image float to the left in order to move the text to the right of each image. The correct border was chosen for the solid borders, as well as the double borders. The copyright was also changed to red font and the font was made smaller just like the example. The track list's list style type were also changed from disc to square. The record labels were changed from normal font to red and italic. Lastly, I changed the line height on the paragraph's, as the spacing was too close.